webapp/infrastructure/requirements.txt
# Modal deployment requirements
fastapi
uvicorn[standard]
pydantic>=2,<3
pydantic-settings
python-docx
docxtpl
openai>=1.30.0
python-multipart
python-dotenv
